variables added to tf.GraphKeys.LOCAL_VARIABLES
experimental at this point. in the process of using
SyncReplicasOptimizer and sharing policy and value parameters
observed master completing initialize_global_variables op and
proceeding with training but workers hanging, waiting for variables
to be initialized (i.e. all variables that were not being shared
via parameter servers). to address this, all non-shared variables
were explicitly added to tf.GraphKeys.LOCAL_VARIABLES so that
once global variable initialization is completed by the master,
workers can continue to initialize the remaining variables with a
local init op.

this is experimental for the reason that it might be more convenient
to programmatically specify which variable collections should be
shared vs. local and add them to the respective collections that way
instead of requiring collections=[tf.GraphKeys.LOCAL_VARIABLES] to
be added to numerous variable declarations throughout the codebase.
unfortunately with the approach I tried for doing this various local
variables were still left in the global collection and thus the
worker local init continued to be blocked.